Implemented Gray's desired order formats (see #2) on both the REST server and the front end.  This required adding some order types to the DB.  Stuff is far from done though.
I implemented factories for most of the objects on the front end, there
has to be a more angular way to do this.

I also brought over the Mayofest stuff to enable posts.  It's weird that
this doesn't just work by default.. (well, maybe posts to PHP are what
doesn't work, but I'd prefer it send it properly than to have to adjust
my server to fiddle with it on receipt.)

Also added in my `config` factory from the mayofest site to carry around
the server url.

I implemented a few buttons, mostly just as demonstrations of how to do
stuff.

The front end now just needs to be able to accomodate different
seasons/turn states.. Such as retreat and supply.

For the backend, I'm still not sure whether I'll be reverting the
Lagragian units tomorrow.  This was a terrible idea.  If I do, I'll also
revert from _retreat/_supply reasons back to turn statuses, where orders
don't get issues unless the retreats can resolve.  This ensures that
state is always consistant, rather than having a consistant state table
and an inconsistant Unit table also saving state)  ... figure it out in
the morning.

In all likelyhood, this won't be ready by 5 tomorrow.

Still to do:
- build sudo util
- use convoy chains
- make sure all orders work
- Implement Create order
- Make sure vacant territories don't get a default point in tallies
- Validate moves based on territories.  Though since the front end does
  this, the priority on the server side is lower (for tomorrow)